backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.17
google-cloud-storage==2.18.2
pydantic-settings==2.6.1
